+++ Core/ClicShopping/Apps/Payment/Stripe/Module/Payment/ST.php
@@ -128,7 +128,7 @@ class ST implements PaymentInterface
         $this->enabled = false;
       }
 
-      $this->sort_order = \defined('CLICSHOPPING_APP_STRIPE_ST_SORT_ORDER') ? CLICSHOPPING_APP_STRIPE_ST_SORT_ORDER : 0;
+      $this->sort_order = \defined('CLICSHOPPING_APP_STRIPE_ST_SORT_ORDER') ? (int)CLICSHOPPING_APP_STRIPE_ST_SORT_ORDER : 0;
     }
   }
