+++ Core/ClicShopping/Apps/Shipping/Item/Module/Shipping/IT.php
@@ -61,7 +61,7 @@ class IT implements ShippingInterface
     $this->code = 'IT';
     $this->title = $this->app->getDef('module_item_title');
     $this->public_title = $this->app->getDef('module_item_public_title');
-    $this->sort_order = \defined('CLICSHOPPING_APP_ITEM_IT_SORT_ORDER') ? CLICSHOPPING_APP_ITEM_IT_SORT_ORDER : 0;
+    $this->sort_order = \defined('CLICSHOPPING_APP_ITEM_IT_SORT_ORDER') ? (int)CLICSHOPPING_APP_ITEM_IT_SORT_ORDER : 0;
 
 // Activation module du paiement selon les groupes B2B
     if ($CLICSHOPPING_Customer->getCustomersGroupID() != 0) {
